third_party/WebKit/LayoutTests/fast/css/usecounter-after.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style id="style"></style>
<script>
'use strict';

test(() => {
    var kHasBeforeOrAfterPseudoElement = 2204; // From UseCounter.h
    var style = document.getElementById('style');
    assert_false(internals.isUseCounted(document, kHasBeforeOrAfterPseudoElement));
	style.textContent = "div::after { color: red; }";
    assert_true(internals.isUseCounted(document, kHasBeforeOrAfterPseudoElement));
}, 'Use of ::after in CSS is use counted.');
</script>
